refactor(my-page): apply code review — glossary names and split children
- `MyDot` / `my-dot.tsx` become `NodeProfile` / `node-profile.tsx`, and the
  view key `"dot"` becomes `"node"`. `CONTEXT.md` puts "dot" on the _Avoid_
  line for **Node**, licensed for **Find your dot**'s copy alone, and this is
  the **Node profile** — a node's appearance. The tab's label still reads
  "My dot", which is the design's copy and stays.
- `Identity`, `ReviewColumn`, `DeleteReviewDialog` and `StatCard` move into
  their own files, matching how `features/collections` splits its named
  children. `my-page.tsx` keeps the data, the tabs and the page's own
  loading/error/signed-out chrome.
- The picture upload holds its local preview until the refreshed `user.me` is
  in the cache. Clearing it first showed the old picture for the length of the
  refetch, and revoking the object URL under a rendered `<img>` blanked the
  avatar.
- One control for the upload: a `<label>` for the file input, rather than a
  button clicking a second, separately-labelled input.
- "1 member found them helpful", not "1 members".
- `personalization-tiers.ts` records that the design contradicts itself on
  tiers 2 and 3: its rendered list pairs 2 with style and 3 with signal, while
  `cc-store.js`'s `TIER_AXES` pairs them the other way round. The rendered list
  wins; the schema is silent.
